Onboarding and Login using Firebase
Created onboarding screens with images as background that shows only once after the app has been installed for the first time.
Also created a login screen that shows for the first time if you haven't logged be logged in before.
and I also edited the movie and tv detail screen.